backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
SQLAlchemy[asyncio]==2.0.38
asyncpg==0.30.0
alembic==1.14.1
python-multipart==0.0.20
redis==5.2.1
httpx==0.28.1
tenacity==9.0.0
opencv-python-headless==4.10.0.84
numpy==2.2.1
pytesseract==0.3.13
PyJWT==2.9.0
email-validator==2.2.0
